Changed type of rating from short to float.

+++ monolith/src/main/kotlin/ru/fmtk/khlystov/culture_code/model/ratings/UserItemRating.kt
@@ -5,7 +5,6 @@ import org.springframework.data.mongodb.core.index.CompoundIndex
 import org.springframework.data.mongodb.core.index.CompoundIndexes
 import org.springframework.data.mongodb.core.mapping.DBRef
 import org.springframework.data.mongodb.core.mapping.Document
-import ru.fmtk.khlystov.culture_code.model.User
 
 @Document(collection = "Ratings_UserItemRating")
 @CompoundIndexes(CompoundIndex(name = "itemTypeAndId", def = "{'itemType' : 1, 'itemId': 1}"))
@@ -13,4 +12,4 @@ data class UserItemRating(@Id val id: String?,
                           @DBRef val userId: String,
                           val itemType: ItemType,
                           val itemId: String,
-                          val rating: Short)
+                          val rating: Float)
